+++ packages/inference/src/index.ts
@@ -1,3 +1,3 @@
 export { HfInference, HfInferenceEndpoint } from "./HfInference";
 export type * from "./types";
-export * from "./lib";
+export * from "./tasks";
