+++ packages/inference/src/tasks/imageClassification.ts
