+++ packages/inference/src/tasks/imageSegmentation.ts
