+++ packages/inference/src/tasks/imageToText.ts
