+++ packages/inference/src/tasks/objectDetection.ts
