+++ packages/inference/src/tasks/textGenerationStream.ts
