+++ packages/inference/src/tasks/textToImage.ts
